<深开鸿><merge><接通speke和client接收数据处理以及调试>

+++ adapter/include/iotc_sle_server.h
@@ -524,7 +524,7 @@ typedef struct {
     IotcAdptSleSsapsChar *character;
     uint8_t charNum;
     int32_t svcHandle;
-    int32_t serverId;
+    uint8_t serverId;
 } IotcAdptSleSsapService;
